Configurando Swagger UI e começando a implementação da autentificação

+++ Models/Book.cs
@@ -1,8 +1,11 @@
+using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
+
 namespace BackendTest.Models
 {
     public class Book
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public string Genre { get; set; }
